+++ src/Zazz.Core/Models/Data/UserDetail.cs
@@ -26,7 +26,7 @@ namespace Zazz.Core.Models.Data
         [ForeignKey("SchoolId")]
         public virtual School School { get; set; }
 
-        public short? SchoolId { get; set; }
+        public int? SchoolId { get; set; }
 
         [ForeignKey("MajorId")]
         public virtual Major Major { get; set; }
